+++ packages/core/src/icon/icon.stories.ts
@@ -11,5 +11,8 @@ export const API: StoryObj = {
     <sl-icon name="star" size="sm"></sl-icon>
     <sl-icon name="star" size="md"></sl-icon>
     <sl-icon name="star" size="lg"></sl-icon>
+    <sl-icon name="check" size="sm"></sl-icon>
+    <sl-icon name="check" size="md"></sl-icon>
+    <sl-icon name="check" size="lg"></sl-icon>
   `
 };