+++ packages/grid/src/locales/nl.ts
@@ -1,0 +1,9 @@
+// Do not modify this file by hand!
+// Re-generate this file by running lit-localize
+
+/* eslint-disable no-irregular-whitespace */
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+export const templates = {
+  sd0b6082239185272: `Select all`
+};