+++ packages/tokens/lib/formats.js
@@ -29,7 +29,10 @@ const tokenToCss = (dictionary, token, options = { prefix: '  ' }) => {
     return `${options.prefix}--${token.name}: ${r} ${g} ${b};`;
   } else if (typeof value === 'object' && token.type === 'typography') {
     const [fontFamily, fontSize, fontWeight, lineHeight] = ['fontFamily', 'fontSize', 'fontWeight', 'lineHeight'].map(
-      attr => replaceReferences(dictionary, token.original.value[attr], value[attr].toString())
+      attr => {
+        if(!value[attr]){console.log('tokenToCss',{dictionary, value, attr, original: token.original.value});}
+        return replaceReferences(dictionary, token.original.value[attr], value[attr].toString());
+      }
     );
 
     return `${options.prefix}--${token.name}: ${fontWeight} ${fontSize}/${lineHeight} ${fontFamily};`;